first part of resturcture ui classes
