+++ src/intelli/gui/ui/packageui.cpp
@@ -7,21 +7,17 @@
  *  Author: Marius Bröcker <marius.broecker@dlr.de>
  */
 
-#include "intelli/gui/packageui.h"
+#include "intelli/gui/ui/packageui.h"
 
-#include "intelli/graph.h"
 #include "intelli/graphcategory.h"
 #include "intelli/package.h"
+#include "intelli/private/utils.h"
 
 #include "gt_icons.h"
-#include "gt_regexp.h"
-#include "gt_qtutilities.h"
-#include "gt_inputdialog.h"
-#include "gt_datamodel.h"
 
 #include <QJsonObject>
 #include <QJsonDocument>
-#include <QRegExpValidator>
+
 #include <QFileInfo>
 #include <QFile>
 
@@ -36,31 +32,7 @@ inline void setObjectName(Node& obj, QString const& name)
     obj.setCaption(name);
 }
 
-template <typename T>
-inline void addNamedChild(GtObject& obj)
-{
-    GtInputDialog dialog{GtInputDialog::TextInput};
-    dialog.setWindowTitle(QObject::tr("Name new Object"));
-    dialog.setWindowIcon(gt::gui::icon::rename());
-    dialog.setLabelText(QObject::tr("Enter a name for the new object."));
 
-    dialog.setTextValidator(new QRegExpValidator{
-        gt::re::onlyLettersAndNumbersAndSpace()
-    });
-
-    if (!dialog.exec()) return;
-
-    QString text = dialog.textValue();
-    if (text.isEmpty()) return;
-
-    auto child = std::make_unique<T>();
-    setObjectName(*child, gt::makeUniqueName(text, obj));
-
-    if (gtDataModel->appendChild(child.get(), &obj).isValid())
-    {
-        child.release();
-    }
-}
 
 PackageUI::PackageUI()
 {
@@ -69,10 +41,6 @@ PackageUI::PackageUI()
     addSingleAction(tr("Add Category"), addNodeCategory)
         .setIcon(gt::gui::icon::add())
         .setVisibilityMethod(isPackageObject);
-
-    addSingleAction(tr("Add Intelli Graph"), addNodeGraph)
-        .setIcon(gt::gui::icon::add())
-        .setVisibilityMethod(isCategoryObject);
 }
 
 QIcon
@@ -90,7 +58,7 @@ PackageUI::addNodeCategory(GtObject* obj)
 {
     if (!isPackageObject(obj)) return;
     
-    addNamedChild<GraphCategory>(*obj);
+    utils::addNamedChild<GraphCategory>(*obj);
 }
 
 void
@@ -98,13 +66,13 @@ PackageUI::addNodeGraph(GtObject* obj)
 {
     if (!isCategoryObject(obj)) return;
     
-    addNamedChild<Graph>(*obj);
+    utils::addNamedChild<Graph>(*obj);
 }
 
 bool
 PackageUI::isCategoryObject(GtObject* obj)
 {
-    return qobject_cast<GraphCategory*>(obj);
+    return false;
 }
 
 bool
