+++ src/intelli/private/utils.h
@@ -17,8 +17,15 @@
 
 #include <gt_state.h>
 #include <gt_statehandler.h>
+#include <gt_inputdialog.h>
+#include <gt_icons.h>
+#include <gt_qtutilities.h>
+#include <gt_datamodel.h>
+#include <gt_regexp.h>
 
 #include <gt_logstream.h>
+
+#include <QRegExpValidator>
 
 #define GT_INTELLI_PROFILE() \
 intelli::Profiler profiler__{__FUNCTION__}; (void)profiler__;
@@ -208,6 +215,32 @@ setupState(GtObject& guardian,
     );
 
     return SetupStateHelper<GetValue>{state, std::move(getValue)};
+}
+
+template <typename T>
+inline void addNamedChild(GtObject& obj)
+{
+    GtInputDialog dialog{GtInputDialog::TextInput};
+    dialog.setWindowTitle(QObject::tr("Name new Object"));
+    dialog.setWindowIcon(gt::gui::icon::rename());
+    dialog.setLabelText(QObject::tr("Enter a name for the new object."));
+
+    dialog.setTextValidator(new QRegExpValidator{
+        gt::re::onlyLettersAndNumbersAndSpace()
+    });
+
+    if (!dialog.exec()) return;
+
+    QString text = dialog.textValue();
+    if (text.isEmpty()) return;
+
+    auto child = std::make_unique<T>();
+    setObjectName(*child, gt::makeUniqueName(text, obj));
+
+    if (gtDataModel->appendChild(child.get(), &obj).isValid())
+    {
+        child.release();
+    }
 }
 
 } // namespace utils
